requirements.txt
langchain-core>=0.1.0
langchain-community>=0.0.1
langgraph>=0.0.1
tavily-python>=0.0.1
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8
html2text
tqdm
transformers
newspaper3k==0.2.8
lxml[html_clean]
faiss-cpu
torch>=1.11.0
sentence-transformers==4.1.0
numpy==1.24.3